Add astype(float64)

+++ dataset.py
@@ -239,12 +239,8 @@ class ImageFolder(data.Dataset):
         filename1 = self.imgs[index]
         filename2 = self.imgs[index + 1]
 
-        # if self.v_compress:
-        img1, fn1 = self.get_frame_data(filename1)
-        img2, fn2 = self.get_frame_data(filename2)
-        # else:
-        # img, main_fn = self.get_frame_data(filename)
-
+        img1, fn1 = self.get_frame_data(filename1).astype(np.float64)
+        img2, fn2 = self.get_frame_data(filename2).astype(np.float64)
         img = np.concatenate((img1, img2), axis=2)
 
         assert img.shape[2] == 6
